# seaf-archtool-core/public/documentation/docs/manual/backend/api.yaml
...
openapi: 3.0.3
info:
  version: 1.0.0
  title: SEAF.ArchTool API
  description: >
    Данный документ актуален для развертывания в режимах Client-Server и Cluster.
    Здесь предоставляются интеграционные API, которые вы можете использовать для 
    встраивания SEAF.ArchTool в свои производственные процессы.
    Подробнее тут https://gitverse.ru/seafteam/seaf-archtool-core/content/master/public/documentation/docs/manual/config/deployment.md
servers:
  - url: https://{URL}
    description: API server secure
    variables:
      URL:
        default: localhost:3030
        description: Сервер https для запроса
  - url: http://{URL}
    description: API server insecure
    variables:
      URL:
        default: localhost:3030
        description: Сервер http для запроса
...
paths:
...
  /seaf-core/api/core/storage/reload:
    put:
      tags:
        - DataLake
      summary: Перезагрузка манифестов в кластере
      description: |
        **В режиме backend**:
          - перезагружает манифест, дополнительные параметры не нужны и не учитываются в обработке, необходимо передать только secret
        
        **В режиме cluster**:
          - работает по указанным ниже параметрам. Умеет:
            - перезагружать выбранные манифесты
            - перезагружать все манифесты
            - обнаруживать ссылки на новые манифесты в root манифесте
        Перезагружает указанные манифесты или все манифесты в кластере.
        Может использоваться для принудительного обновления кэша.
        
        В случае возникновения ошибки, можно поискать в логах по значению 'trace-id-123' из ответа с ошибкой
      operationId: reloadClusterStorage
      parameters:
        - name: secret
          in: query
          description: secret для доступа к запросу. Указывается в .env переменной VUE_APP_DOCHUB_RELOAD_SECRET
          required: true
          schema:
            type: string
          example: "your-reload-secret-key"

        - name: reloadRoot
          in: query
          description: |
            Если true - запускает перезагрузку манифестов для подтягивания изменений из root манифеста. 
            Перезагрузка может запуститься если ранее она уже была запрошена (другим запросом) и не успела обработаться
            При этом параметры alias и mode игнорируются если переданы.
          schema:
            type: string
            enum: ["true", "false"]
            default: "false"
          example: "false"

        - name: mode
          in: query
          description: |
            Режим работы с параметром alias.
            - `alias` (по умолчанию) - alias трактуются как имена манифестов
            - `permission` - alias трактуются как разрешения, которые конвертируются в имена манифестов через root manifest
          schema:
            type: string
            enum: [alias, permission]
            default: alias
          example: "alias"

        - name: alias
          in: query
          description: |
            Имена манифестов для перезагрузки.
            
            **Формат передачи:**
            - Для одного значения: `?alias=manifest1`
            - Для нескольких значений: `?alias=manifest1&alias=manifest2&alias=manifest3`
            
            Если параметр не указан - перезагружаются все доступные манифесты.
            
            **Важно:** При `mode=permission` этот параметр содержит имена разрешений,
            а не имена манифестов.
            
            **Формат при `mode=permission` не отличается, но в alias ожидаются значения permission :**
            - Для одного значения: `?alias=permission1`
            - Для нескольких значений: `?alias=permission1&alias=permission2&alias=permission3`
          style: form
          explode: true
          schema:
            type: array
            items:
              type: string
          example: ["manifest1"]

      responses:
        '200':
          description: Запрос успешно обработан
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "success"
                example: {
                  "message": "success"
                }

        '400':
          description: |
            Неверные параметры запроса
            В режиме backend если что-то пошло не так, тоже возвращается 400
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "trace-id-123: bad params 'alias' (nir)"

        '403':
          description: Неверный секретный ключ
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "trace-id-123: Error reload secret is not valid [wrong-secret]"

        '500':
          description: Внутренняя ошибка сервера или сервер не готов
          headers:
            Retry-After:
              description: Время в секундах, через которое можно повторить запрос
              schema:
                type: string
                example: "60"
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "trace-id-123: server not ready (aliases empty in storage)"
...